refactor: login and logout
