+++ src/routes/auth.route.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import AuthController from '../modules/student/AuthController';
+import { requiredAuthentication } from '../middlewares/authentication/requiredAuthentication';
+
+export const authRouter = Router();
+
+authRouter.post('/login', AuthController.login);
+authRouter.post('/logout', requiredAuthentication, AuthController.logout);
