+++ src/routes/index.ts
@@ -2,14 +2,12 @@ import { Router } from 'express';
 import { materialRouter } from './material.route';
 import { rootRouter } from './root.routes';
 import { studentRouter } from './student.route';
-import { loginRouter } from './login.route';
 import { categoryRouter } from './category.route';
-import { logoutRouter } from './logout.route';
+import { authRouter } from './auth.route';
 
 export const routes = Router();
 
-routes.use('/login', loginRouter);
-routes.use('/logout', logoutRouter);
+routes.use('/auth', authRouter);
 routes.use('/student', studentRouter);
 routes.use('/category', categoryRouter);
 routes.use('/material', materialRouter);
