--- src/routes/login.route.ts
@@ -1,6 +1,0 @@
-import { Router } from 'express';
-import LoginController from '../modules/student/LoginController';
-
-export const loginRouter = Router();
-
-loginRouter.post('/', LoginController.login);
